+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed PBR bug when multilayer: false is defined
 - Fixed upgrade-shader not generating mipmaps
+- Fixed PBR bug with duplicate texture sets
 
 ## [0.5.7] - 2025-09-05